p4i integration Integrate 264909 from 4.3 to main: Fixes:    crash caused by a trucated cache key value in the gradient cache Task:     163690 RevBy:    Prasanth
[git-p4: depot-paths = "//depot/qt/main/": change = 264910]

+++ src/opengl/qpaintengine_opengl.cpp
@@ -979,14 +979,13 @@ protected:
     GLuint addCacheElement(quint64 hash_val, const QGradientStops &stops, qreal opacity) {
         if (cache.size() == maxCacheSize()) {
             int elem_to_remove = qrand() % maxCacheSize();
-            uint key = cache.keys()[elem_to_remove];
+            quint64 key = cache.keys()[elem_to_remove];
 
             // need to call glDeleteTextures on each removed cache entry:
             QGLGradientColorTableHash::const_iterator it = cache.constFind(key);
             do {
                 glDeleteTextures(1, &it.value().texId);
             } while (++it != cache.constEnd() && it.key() == key);
-
             cache.remove(key); // may remove more than 1, but OK
         }
         CacheInfo cache_entry(stops, opacity);
